Show add to wish list if user is not logged on

+++ app/helpers/lessons_helper.rb
@@ -113,14 +113,12 @@ module LessonsHelper
   end
 
   def button_to_wish(lesson)
-    if current_user
-      if lesson.instructed_by?(current_user) or lesson.owned_by?(current_user)
-        return
-      elsif current_user.on_wish_list?(lesson)
-        button_to "Remove from Wish List", wish_lists_path(:id => lesson), :method => :delete, :disable_with => translate('general.disable_with')
-      else
-        button_to "Add to Wish List", wish_lists_path(:id => lesson), :method => :post, :disable_with => translate('general.disable_with')
-      end
+    if lesson.instructed_by?(current_user) or lesson.owned_by?(current_user)
+      return
+    elsif current_user and current_user.on_wish_list?(lesson)
+      button_to "Remove from Wish List", wish_lists_path(:id => lesson), :method => :delete, :disable_with => translate('general.disable_with')
+    else
+      button_to "Add to Wish List", wish_lists_path(:id => lesson), :method => :post, :disable_with => translate('general.disable_with')
     end
   end
